don't cut a symbol with the envelope; use a gaussian filter
If we cut the symbol with the envelope, we might severely limit the
audibility of a dit.  Implement trailing samples for the sloped parts
that get merged with following symbol.

Switch to a gaussian envelope that minimizes key clicks.
Based on <http://fermi.la.asu.edu/w9cf/articles/click/index.html>